+++ test/Wallet/walletGabaim.t.sol
@@ -1,5 +1,6 @@
 // SPDX-License-Identifier: MIT
 pragma solidity ^0.8.20;
+
 import "foundry-huff/HuffDeployer.sol";
 import "forge-std/Test.sol";
 import "forge-std/console.sol";
@@ -18,6 +19,7 @@ contract TestWalletGabaim is Test {
         vm.deal(address(this), 100);
         // מעבירה לארנק חלק מהכסף שנמצא בכתובת
     }
+
     function testReceive() public {
         uint256 balance = walletG.getValue();
         uint256 amount = 50;
@@ -32,11 +34,13 @@ contract TestWalletGabaim is Test {
         vm.expectRevert("you are owner");
         walletG.changeOwners(newCollector, oldCollector);
     }
+
     function testChangeOwner() public {
         address oldCollector = 0x5B38Da6a701c568545dCfcB03FcB875f56beddC4;
         address newCollector = 0x9876543210987654321098765432109876543210;
         walletG.changeOwners(newCollector, oldCollector);
     }
+
     function testwithDrawIsnOwner() public {
         uint256 amountWithDraw = 50;
         vm.startPrank(noOwnerAddress);
@@ -45,12 +49,13 @@ contract TestWalletGabaim is Test {
         vm.stopPrank();
         // assertEq(walletG.getValue(), balance);
     }
+
     function testwithDrawIsOwner() public {
         //  payable(address(walletG)).transfer(200);
         vm.deal(address(walletG), 200);
         //  payable(address(walletG)).transfer(200);
         uint256 amountWithDraw = 50;
-        uint balance = address(walletG).balance;
+        uint256 balance = address(walletG).balance;
         vm.startPrank(ownerAddress);
         walletG.withDraw(amountWithDraw);
         assertEq(walletG.getValue(), balance - amountWithDraw);
